Move NumericsConfig into ProcessLib.
